Adds an unoptimized in-memory ledger to snarkvm-ledger
