+++ .integration/src/lib.rs
@@ -14,69 +14,26 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use snarkvm_utilities::{FromBytes, ToBytes};
-
-use std::io::{Read, Result as IoResult, Write};
-
-pub const COL_META: u32 = 0; // MISC Values
-pub const COL_BLOCK_HEADER: u32 = 1; // Block hash -> block header
-pub const COL_BLOCK_TRANSACTIONS: u32 = 2; // Block hash -> block transactions
-pub const COL_BLOCK_LOCATOR: u32 = 3; // Block num -> block hash && block hash -> block num
-pub const COL_TRANSACTION_LOCATION: u32 = 4; // Transaction Hash -> (block hash and index)
-pub const COL_COMMITMENT: u32 = 5; // Commitment -> index
-pub const COL_SERIAL_NUMBER: u32 = 6; // SN -> index
-pub const COL_DIGEST: u32 = 8; // Ledger digest -> index
-pub const COL_RECORDS: u32 = 9; // commitment -> record bytes
-pub const COL_CHILD_HASHES: u32 = 10; // block hash -> vector of potential child hashes
-pub const NUM_COLS: u32 = 11;
-
-pub const KEY_BEST_BLOCK_NUMBER: &str = "BEST_BLOCK_NUMBER";
-pub const KEY_CURR_CM_INDEX: &str = "CURRENT_CM_INDEX";
-pub const KEY_CURR_SN_INDEX: &str = "CURRENT_SN_INDEX";
-pub const KEY_CURR_DIGEST: &str = "CURRENT_DIGEST";
-
-pub mod ledger;
-pub use ledger::*;
-
-pub mod memdb;
-pub use memdb::*;
-
 pub mod testnet1;
 pub use testnet1::*;
 
 pub mod testnet2;
 pub use testnet2::*;
 
-/// Represents address of certain transaction within block
-#[derive(Debug, PartialEq, Clone)]
-pub struct TransactionLocation {
-    /// Transaction index within the block
-    pub index: u32,
-    /// Block hash
-    pub block_hash: [u8; 32],
+use snarkvm_dpc::prelude::*;
+use snarkvm_ledger::prelude::*;
+
+use rand::Rng;
+
+pub fn random_storage_path() -> String {
+    let random_path: usize = rand::thread_rng().gen();
+    format!("./test_db-{}", random_path)
 }
 
-impl ToBytes for TransactionLocation {
-    #[inline]
-    fn write_le<W: Write>(&self, mut writer: W) -> IoResult<()> {
-        self.index.write_le(&mut writer)?;
-        self.block_hash.write_le(&mut writer)
-    }
-}
+/// Initializes a test ledger given a genesis block.
+pub fn initialize_test_blockchain<C: Parameters, S: Storage>(genesis_block: Block<Transaction<C>>) -> Ledger<C, S> {
+    let mut path = std::env::temp_dir();
+    path.push(random_storage_path());
 
-impl FromBytes for TransactionLocation {
-    #[inline]
-    fn read_le<R: Read>(mut reader: R) -> IoResult<Self> {
-        let index: u32 = FromBytes::read_le(&mut reader)?;
-        let block_hash: [u8; 32] = FromBytes::read_le(&mut reader)?;
-
-        Ok(Self { index, block_hash })
-    }
-}
-
-pub fn bytes_to_u32(bytes: &[u8]) -> u32 {
-    let mut num_bytes = [0u8; 4];
-    num_bytes.copy_from_slice(&bytes);
-
-    u32::from_le_bytes(num_bytes)
+    Ledger::new(Some(&path), genesis_block).unwrap()
 }
