+++ .integration/tests/dpc_testnet1.rs
@@ -21,7 +21,7 @@ use snarkvm_algorithms::{
 };
 use snarkvm_curves::bls12_377::{Fq, Fr};
 use snarkvm_dpc::{prelude::*, testnet1::parameters::*};
-use snarkvm_integration::{ledger::*, memdb::MemDb, testnet1::*};
+use snarkvm_integration::{testnet1::*, *};
 use snarkvm_ledger::prelude::*;
 use snarkvm_r1cs::{ConstraintSystem, TestConstraintSystem};
 use snarkvm_utilities::{to_bytes_le, FromBytes, ToBytes};
