+++ ledger/Cargo.toml
@@ -96,6 +96,9 @@ version = "0.4.3"
 [dependencies.once_cell]
 version = "1.8.0"
 
+[dependencies.parking_lot]
+version = "0.11.1"
+
 [dependencies.rand]
 version = "0.8"
