+++ ledger/src/ledger/memdb.rs
@@ -14,8 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::NUM_COLS;
-use snarkvm_ledger::prelude::*;
+use crate::{ledger::NUM_COLS, prelude::*};
 
 use parking_lot::RwLock;
 use std::{collections::HashMap, path::Path};
